test-case-manager/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
httpx==0.25.0
aiohttp==3.9.0
google-genai>=0.2.0
python-multipart==0.0.6
PyJWT==2.8.0
datasets>=2.14.0
requests>=2.31.0


